Create Account, Session, and User tables in Postgres database
This commit introduces the initial SQL migrations that create the Account, Session, and User tables in the Postgres database. It also sets up primary keys, unique indices, and a couple of foreign key constraints. The migration_lock file has also been added, marking the usage of Postgresql as a provider for Prisma.