Add Mailgun integration to support sending email
Updated the application to support sending emails with Mailgun. In the server.js file, Mailgun is now required and used to send a magic link email for passwordless user authentication. The mailgun.js dependency was also added to the package.json and package-lock.json. Session secret setting has been updated to use an environment variable or a default fallback.