Implement passwordless authentication
Removed password-based authentication from the system and switched to a passwordless authentication model, making use of magic links. The changes span across the front-end and back-end, involving several changes to the user registration and login endpoints, update on the helpers utilities, and also modification of related tests. This provides an improved user experience and enhanced security.