Remove email-password tests in server, auth and poll
Tests related to email and password authentication have been removed from the 'server.test.js', 'auth.test.js', and 'poll.test.js' files. This is because the system has been updated to follow a passwordless authentication method, resulting in the removal of the aforementioned tests that dealt with password-based authentication.