Add new columns to User table and refactor authentication process
Added "loginToken" and "tokenExpires" columns to the User table in the Prisma migration file. Adjusted Google OAuth code to better verify user and account information. This change also involved optimizing routes in the 'server.js' file to align with the adjusted authentication method. Also included the 'form-data' package to help with data handling for the new user columns.