Added Net Benefits Calculator feature.
New HTML, JS and CSS files have been created for the Net Benefits Calculator feature in the public directory. The calculator provides a dynamic user interface to input parameters and calculate the net benefits of shifting military spending to healthcare. A new entry to ignore data files has been added to .gitignore as part of this commit.