Update sign-in/up UI text and user image attribute, add PowerShell kill node script
The user interface for sign-in and sign-up was updated for a better branding experience, changing the text to match the fantasy theme of 'Wishonia'. The 'imageUrl' attribute for user has been renamed to 'image' in the codebase for simplicity and readability. A PowerShell script for stopping node process was also added.