Add spending on diseases vs military chart
This commit introduces a new chart that compares the amount of money spent on diseases versus the military. The chart is rendered using the Chart.js library and the relevant data is hardcoded into the Javascript file. A new dependency, 'chartjs-plugin-datalabels', is added to the package.json to enable data labels on the chart. Also, the chart is integrated into the dashboard page.