Remove MDX and related code
The MDX related code and files have been deleted from the application. This includes the deletion of some MDX pages, the removal of MDX from the Next config, and adjustments to the post page to adapt to these changes. These modifications reflect a decision to move away from using MDX for content management in this project.