Update README to reflect project focus on societal wish fulfillment
The README file has been overhauled to reflect the change in the project's concept and scope from global research spending analysis to "Wishonia", an imaginary kingdom focused on optimal allocation of societal resources to maximize citizens' wish fulfillment. The new approach includes the identification and prioritization of societal problems, the proposal of solutions, allocation of resources, the assignment of tasks, monitoring progress and reporting results. The README now also includes a detailed description of the process, as well as an example of how a societal budget might be created. Installation instructions have been updated to reflect the use of `pnpm`.