Add seed script to package.json and update Prisma schema and seed data
The update includes the addition of a "seed" script in package.json for easier database seeding. Also, a "name" field is included in the Prisma schema and some data in the Prisma seed file has been adjusted. This includes a change in the username and the inclusion of actual problem content.