Add customizable nav items to user avatar menu
Updated the UserNavDisplay and UserAccountNav components to support customizable navigation items. The avatarNavItems prop has been added to both components, allowing the navigation options in the user's avatar dropdown menu to be configured. If no custom nav items are provided, the default values from the avatarNav set in the config are used.