Add components for Global Problems List and Pie Chart
This commit includes the addition of new components for the Global Problems List and the Global Problems Pie Chart. These components fetch global problem data from the API, display this data in a sortable table list, as well as a responsive pie chart. The table columns and the pie chart are defined in separate modules for reusability and cleanliness of code.