Add postVoteData function and refactor code related to vote data
A new function, postVoteData, was added to handle the posting of vote data to the '/api/vote' endpoint. In addition, some codes related to polling and voting were refactored. For instance, the getAverageWarPercentageDesired function was moved from "poll.ts" to "warVsCures.ts", and the function call in the dashboard.ts was updated accordingly.