Update data seeding and allocation process for Wishing Wells and Problems
This commit includes updates to the seeding process for Wishing Wells and Global Problems. The most significant alterations are the change in data return for the `seedWishingWells` function and the introduction of a system to prevent multiple allocations for the same pair in both `seedWishingWellPairAllocations` and `seedGlobalProblemPairAllocations`. Furthermore, a more efficient combination generation method is implemented in the `generateAllWishingWellPairs` function, and the names in the `GlobalProblem` and `WishingWell` models in the schema are enforced to be unique.