Add image compression script and update README
A new script, compressImages.js, has been added to compress images in specified directories. This script traverses the directories and compresses any image files it finds. Additionally, the README has been updated with new details about Wishocracy functions and the concept of Digital Twins in the simulated World of Wishonia.