Add image compression configuration options
The compressionImages.js script has been updated to provide greater flexibility with image compression configuration. It now allows setting the output quality and the output format for the images. Furthermore, the logic for converting PNG images to JPG has also been added to enable the alteration of the image file format post-compression.