Refactor metadata and markdown generation, add image check
The commit includes refactoring of metadata and markdown file generation and introduces a check for existing images. It also moves "wishingWellGenerator.js" to the "lib" directory and renames it "wishingWellMarkdownGenerator.ts". A new generator for global problems was added. Changes in the metadata generator include the ability to handle both "title" and "name" attributes for posts. The image generator now has a check to avoid unnecessary image generation.