Add "Docs" and "Github" links to navigation
Two new navigation items, "Docs" and "Github", were added to the top navigation and the footer. These additions provide users with new ways to access the documentation and the Github project page directly from the application. A new 'book' icon was also added to the icon list to represent the "Docs" link.