Add tests for Github loader queue
A new test suite has been added for the Github load queue. This includes a test to validate the loading and processing of a GitHub repository. The changes also encompass the setup and takedown of a mock Prisma client as well as the configuration of necessary parameters for the test.