Add functionality to upload images to Vercel
A new module 'imageUploader.ts' has been added in the lib directory. This module introduces two new functions - 'uploadImageToVercel' and 'vercelImageExists'. The former handles the task of uploading an image to Vercel using the put method from @vercel/blob, while the latter checks if a certain image already exists in the Vercel CDN.