Add solution voting and average results functionality
A new component "GlobalProblemSolutionsVoteAndSolutionsList" has been created, which includes the functionality for users to vote on solutions for global problems. The solutions' names are now extracted properly and displayed in the list. Additionally, the average allocation of resources towards each solution is shown. A "Solutions" button has also been added next to each global problem in the list, directing users to the corresponding solutions page.