Add getRedisClient function in Redis utility
A new function, getRedisClient, is introduced in the redis.ts file to centralize the creation of a Redis client. This change will make the client generation process more consistent and manageable throughout the project. The getRedisModelCache function is also updated to use this new function instead of creating the client directly.