Add "admin" field to ExtendedUser type
The "admin" field has been added to the ExtendedUser type. This field encapsulates whether or not a user has admin privileges. Now, in the session, the cast to ExtendedUser is also done, and the "admin" property is also added in the library auth. In GlobalProblemSolutionsList, the user type is changed to ExtendedUser and an additional admin check is added.