Add username to ExtendedUser interface
In the ExtendedUser interface, an optional 'username' field has been added. This change was necessary to update the saveReferrerUserId function in vote/route.ts, which now uses ExtendedUser in its parameters instead of a custom User object. As a result, type consistency across the application is reinforced and the code is now cleaner and more manageable.