Implement major overhaul of Agent and related tables
This commit extensively modifies the structure of the Agent table, adding new fields and updating others. It also redefines the related tables by dropping many of them such as AgentApiHistory and AgentDocument, and replacing them with newly designed tables that are better aligned with our updated data model. Changes in this commit lay the groundwork for the next phase of feature development.