Disable Sentry and correct Prisma client imports
The updates to the Sentry initialization now skip the process when the SENTRY_AUTH_TOKEN environment variable isn't set. Doing this can prevent potential crashes during startup in the absence of the token. Also, corrected the import path for Prisma client in multiple files to ensure the correct functioning of Prisma operations.