Update landing page interactivity and fix sever/client render conflict
The landing page content has been heavily updated with interactive and descriptive sentences. Also, a mechanism has been added to dynamically render content based on whether the client has been fully loaded or not. This ensures a smooth experience for the user as content loading is better handled, preventing blank or broken interfaces in the initial rendering phase.